requirements.txt
numpy
matplotlib
scikit-image
Pillow
tensorflow
lime